Refactoring CachedFiles to allow for multiple 'tabs'
