+++ frontend/deps/Styles.css
@@ -351,3 +351,7 @@ li.tree-leaf {
 .colormap-apply-menu.dropdown-menu {
     background-color: #fff /* Reset the background color set by another rule */
 }
+
+.variable-tabs {
+    
+}
