+++ frontend/src/js/components/modals/CachedFiles/RenameVariable.jsx
@@ -1,0 +1,48 @@
+import React, { Component } from 'react'
+import { Modal, Button, FormGroup, FormControl, ControlLabel } from 'react-bootstrap';
+import Dialog from 'react-bootstrap-dialog';
+import FileExplorer from '../FileExplorer/FileExplorer.jsx';
+
+class RenameVariable extends Component{
+    constructor(props){
+        super(props)
+    }
+    
+    render(){
+        return(
+            <div>
+                <Modal show={this.state.showRedefineVariableModal}>
+                    <Modal.Header>
+                        <Modal.Title id="contained-modal-title-sm">Rename variable</Modal.Title>
+                    </Modal.Header>
+                    <Modal.Body>
+                        <FormGroup>
+                            <ControlLabel>Variable name</ControlLabel>
+                            <FormControl
+                                type="text"
+                                value={this.state.temporaryRedefinedVariableName}
+                                onChange={(e) => this.setState({ temporaryRedefinedVariableName: e.target.value })} />
+                        </FormGroup>
+                    </Modal.Body>
+                    <Modal.Footer>
+                        <Button bsStyle="primary" bsSize="small"
+                            onClick={() => {
+                                if (this.state.temporaryRedefinedVariableName) {
+                                    this.doneRedefineVariable();
+                                    this.setState({
+                                        redefinedVariableName: this.state.temporaryRedefinedVariableName,
+                                        showRedefineVariableModal: false
+                                    });
+                                }
+                            }}>Confirm</Button>
+                        <Button bsSize="small" onClick={() => this.setState({ showRedefineVariableModal: false })}>Close</Button>
+                    </Modal.Footer>
+                </Modal>
+                <Dialog ref="dialog" />
+                {this.state.showFileExplorer &&
+                    <FileExplorer show={true} onTryClose={() => this.handleFileExplorerTryClose()} onFileSelected={(file) => this.handleFileSelected(file)} />}
+                }
+            </div>
+        )
+    }
+}
